Change two-command build into a single command recursive one
The problem with the two-command build was that it was awkward for ExternalProject_Add() to make use of it. Change it so that the CMake configure step builds the host tools itself by generating  'host' directory inside the chosen build directory and then building the host tools in there.

Note that for this to work the compiler chosen must be squashed so that the child build doesn't inherit the cross compiler via $CC.